Camera + Pinhole Implemented. Fixes to existing headers

+++ raytracer/world/World.cpp
@@ -39,7 +39,7 @@ ShadeInfo World::hit_objects(const Ray &ray)
     float t = 0.0;
 
 
-    for (int i = 0; i < geometry.size(); i++)
+    for (int i = 0; i < (int) geometry.size(); i++)
     {
         bool hit_flag = geometry[i]->hit(ray, t, tmp);
